Fix for no DKIM public key found due to wrong domain name.
Apparently, OpenDKIM uses the hostname declared in the
KeyTable for signing emails.

Here's a rough explanation of what the whole line does:

    alias hostname:selector:/path/to/private/key

/path/to/private/keyWhere:

/path/to/private/keyalias - corresponds to the name declared in
the SigningTable to indicate which entry to use for signing emails.

hostname - used by OpenDKIM when signing emails. Entry here
will be used as the value for the "d=" field.

selector - not sure what this really is. This is set when creating
an OpenDKIM keypair.

/path/to/private/key - Private key to be used

Signed-off-by: John Paul Alcala <jp@jpalcala.com>