PCI DSS compliant SSL configuration for Postfix and Dovecot
This provides support for the following protocols:

* TLS v1.0
* TLS v1.1
* TLS v1.2

The default cipher suite has been made explicit instead of implicit,
since there are existing ciphers that cannot be excluded via a
selector/family (e.g. CBC-related ciphers).

Also, DH param is now used for Postfix. Usage of an external
DH param file is not supported in this version of Dovecot (2.2.9).

Signed-off-by: John Paul Alcala <jp@jpalcala.com>