Removed all the wait conditions to ensure proper shutdown.
I may be hitting a race condition when the process is killed
and the pid file is cleaned up before it can be read by the
wait command.

I could fix this by storing the PID on a variable, but seeing
that there is little value to this, I'd rather just remove it.
It will produce annoying messages during shutdown, but who
cares?

Signed-off-by: John Paul Alcala <jp@jpalcala.com>